+++ src/joedb/journal/Stream_File.h
@@ -15,14 +15,14 @@ namespace joedb
    std::streambuf &streambuf;
 
   protected:
-   size_t raw_read(char *buffer, size_t size) final;
-   void raw_write(const char *buffer, size_t size) final;
-   void raw_seek(int64_t offset) final;
+   size_t raw_read(char *buffer, size_t size) override;
+   void raw_write(const char *buffer, size_t size) override;
+   void raw_seek(int64_t offset) override;
 
   public:
    Stream_File(std::streambuf &streambuf, Open_Mode mode);
 
-   int64_t get_size() const final;
+   int64_t get_size() const override;
 
    ~Stream_File() override;
  };